requirement.txt
langchain
langchain-core
langchain-openai
openai
PyMuPDF
whisper
nltk
numpy
scikit-learn
fastapi
uvicorn
unstructured
python-pptx
docx2txt
python-dotenv